Moved SavePointTask to the savepoint package
